requirements.txt
albumentations==0.4.3
opencv-python
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
torchmetrics==0.6
omegaconf>=2.3.0
test-tube>=0.7.5
streamlit>=0.73.1
einops>=0.6.0
transformers>=4.30.0
webdataset==0.2.5
open-clip-torch==2.7.0
gradio==3.13.2
kornia==0.6
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.3.0
tqdm>=4.65.0
safetensors>=0.3.1
-e .
